* development: (69 commits)
  build files + hide fb and insta
  fixed: CBI-LARAVEL-19
  updated youtube
  Added youtube login (in progress)
  build file
  update dashboard table settings
  build file
  build file + css fix
  user days graphs
  Updated instagram structure
  redirected facebook and displayed facebook popup
  updated facebook fields
  update graphs
  update graphs
  build file + css fix
  build file + css fix
  build file + css fix
  build file
  fixed ui bugs add button on apps
  build files
  ...

# Conflicts:
#	public/js/UI.js
#	public/mix-manifest.json

+++ resources/js/react/UserInterface/components/settings/Accounts.js
@@ -625,6 +625,7 @@ export default class Accounts extends React.Component {
             let googleAccounts = this.state.googleAccounts;
             googleAccounts = googleAccounts.filter(ga => ga.id != id);
             this.setState({isBusy: false, googleAccounts: googleAccounts})
+            (this.props.reloadUser)();
         }, (err) => {
             this.setState({isBusy: false, errors: (err.response).data});
         }).catch(err => {
@@ -634,13 +635,19 @@ export default class Accounts extends React.Component {
 
     handleFacebookDelete(id) {
         this.setState({isBusy: true});
+        if(confirm('Disconnecting the account, will remove and delete all the annotations related to this account, are you sure?'))
         HttpClient.delete(`/settings/facebook-account/${id}`).then(resp => {
-            Toast.fire({
-                icon: 'success',
-                title: "Account removed.",
-            });
             let facebookAccounts = this.state.facebookAccounts;
             facebookAccounts = facebookAccounts.filter(ga => ga.id != id);
+            if(this.state.facebookAccounts.length > 1) {
+                Toast.fire({
+                    icon: 'success',
+                    title: "Account removed.",
+                });
+            } else {
+                this.updateUserService('is_ds_facebook_tracking_enabled');
+                (this.props.reloadUser)();
+            }
             this.setState({isBusy: false, facebookAccounts: facebookAccounts})
         }, (err) => {
             this.setState({isBusy: false, errors: (err.response).data});
